Update byte receive range

+++ capstone-backend/internal/role/model.go
@@ -1,7 +1,7 @@
 package role
 
 type Role struct {
-	ID   int32   `db:"role_id"   json:"id"`
+	ID   int64   `db:"role_id"   json:"id"`
 	Name string  `db:"role_name" json:"name"`
 	Desc *string `db:"role_desc" json:"desc,omitempty"`
 }
